# Conflicts:
#	build.gradle

+++ src/main/java/com/sparta/quizdemo/comment/service/CommentService.java
@@ -7,6 +7,7 @@ import com.sparta.quizdemo.comment.entity.Comment;
 import com.sparta.quizdemo.comment.repository.CommentRepository;
 import com.sparta.quizdemo.common.dto.ApiResponseDto;
 import com.sparta.quizdemo.user.entity.User;
+import com.sparta.quizdemo.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -25,6 +26,7 @@ import java.util.List;
 @Transactional(readOnly = true)
 public class CommentService {
     private final CommentRepository commentRepository;
+    private final UserRepository userRepository;
 
     @Transactional //댓글 생성
     public ResponseEntity<ApiResponseDto> createComment(CommentRequestDto commentRequestDto, User user) {
@@ -52,24 +54,36 @@ public class CommentService {
     @Transactional //댓글 수정
     public ResponseEntity<ApiResponseDto> updateComment(Long id, CommentRequestDto commentRequestDto, User user) {
         try {
-            //예외가 발생하지않으면 200 상태코드와 메시지 반환
             Comment comment = commentRepository.findById(id)
                     .orElseThrow((() -> (new IllegalArgumentException("댓글이 존재하지 않습니다."))));
-            comment.update(commentRequestDto, user);
-            ApiResponseDto apiResponseDto = new ApiResponseDto("댓글을 수정했습니다.", HttpStatus.OK.value());
-            return new ResponseEntity<>(apiResponseDto, HttpStatus.OK);
+            if (!comment.getUser().getId().equals(user.getId())) {
+                throw new RuntimeException("작성자만 수정 가능합니다.");
+            } else {
+                comment.update(commentRequestDto, user);
+                //예외가 발생하지않으면 200 상태코드와 메시지 반환
+                ApiResponseDto apiResponseDto = new ApiResponseDto("댓글을 수정했습니다.", HttpStatus.OK.value());
+                return new ResponseEntity<>(apiResponseDto, HttpStatus.OK);
+            }
         } catch (Exception e) {
             //예외가 발생하면 예외처리된 원인과 상태코드 400반환
             return ResponseEntity.badRequest().body(new ApiResponseDto(e.getMessage(), HttpStatus.BAD_REQUEST.value()));
         }
     }
 
-    public ResponseEntity<ApiResponseDto> deleteComment(Long id) { // 댓글 삭제
+    @Transactional
+    public ResponseEntity<ApiResponseDto> delete_Comment(Long id, User user) { // 댓글 삭제
+        Comment comment = commentRepository.findById(id)
+                .orElseThrow((() -> (new IllegalArgumentException("댓글이 존재하지 않습니다."))));
         try {
-            //예외가 발생하지않으면 200 상태코드와 메시지 반환
-            commentRepository.deleteById(id);
-            ApiResponseDto apiResponseDto = new ApiResponseDto("댓글을 삭제했습니다.", HttpStatus.OK.value());
-            return new ResponseEntity<>(apiResponseDto, HttpStatus.OK);
+            if (!comment.getUser().getId().equals(user.getId())) {
+                throw new Exception("작성자만 삭제가 가능합니다.");
+            } else {
+                //예외가 발생하지않으면 200 상태코드와 메시지 반환
+                commentRepository.deleteById(id);
+                ApiResponseDto apiResponseDto = new ApiResponseDto("댓글을 삭제했습니다.", HttpStatus.OK.value());
+                return new ResponseEntity<>(apiResponseDto, HttpStatus.OK);
+            }
+
         } catch (Exception e) {
             //예외가 발생하면 예외처리된 원인과 상태코드 400반환
             return ResponseEntity.badRequest().body(new ApiResponseDto(e.getMessage(), HttpStatus.BAD_REQUEST.value()));
@@ -78,37 +92,42 @@ public class CommentService {
 
 
     @Transactional
-    public ResponseEntity<ApiResponseDto> create_CommentLike(Long id, User user) { //댓글 좋아요 생성
+    public ResponseEntity<ApiResponseDto> create_CommentLike(Long user_id, Long comment_id, User user) { //댓글 좋아요 생성
         try {
-            Comment comment = commentRepository.findById(id)
+            Comment comment = commentRepository.findById(comment_id)
                     .orElseThrow(() -> (new IllegalArgumentException("댓글이 존재하지 않습니다.")));
-            if (comment.isBool() == true) {
-                throw new Exception("이미 좋아요가 있는 존재합니다.");
-            } else {
-                comment.setBool(true); // 사용자가 좋아요 설정: true면 좋아요 생성, false: 좋아요 삭제
+            User find_user_id = userRepository.findUserById(user_id)
+                    .orElseThrow(() -> (new IllegalArgumentException("유저가 존재하지 않습니다.")));
+
+            if (comment.isBool() == false) {
+                find_user_id.setUbool(true);
                 comment.setLikeCnt(comment.getLikeCnt() + 1); // 좋아요를 했으면 좋아요 수 증가
+                ApiResponseDto apiResponseDto = new ApiResponseDto("댓글 좋아요가 생성되었습니다.", HttpStatus.OK.value());
+                return new ResponseEntity<>(apiResponseDto, HttpStatus.OK);
+            } else {
+                return ResponseEntity.badRequest().body(new ApiResponseDto("이미 좋아요가 생성되었습니다.", HttpStatus.BAD_REQUEST.value()));
             }
-            ApiResponseDto apiResponseDto = new ApiResponseDto("댓글 좋아요가 생성되었습니다.", HttpStatus.OK.value());
-            return new ResponseEntity<>(apiResponseDto, HttpStatus.OK);
         } catch (Exception e) {
             return ResponseEntity.badRequest().body(new ApiResponseDto(e.getMessage(), HttpStatus.OK.value()));
         }
     }
 
     @Transactional
-    public ResponseEntity<ApiResponseDto> delete_CommentLike(Long id, User user) { //댓글 좋아요 삭제
+    public ResponseEntity<ApiResponseDto> delete_CommentLike(Long user_id, Long comment_id, User user) { //댓글 좋아요 삭제
         try {
-            Comment comment = commentRepository.findById(id)
+            Comment comment = commentRepository.findById(comment_id)
                     .orElseThrow(() -> (new IllegalArgumentException("댓글이 존재하지 않습니다.")));
-            if (comment.isBool() == false) {
-                throw new Exception("이미 좋아요를 삭제했습니다.");
+            User find_user_id = userRepository.findUserById(user_id)
+                    .orElseThrow(() -> (new IllegalArgumentException("유저가 존재하지 않습니다.")));
+            if (!comment.getUser().getId().equals(user.getId()) || comment_id.equals(comment.getId())) {
+                find_user_id.setUbool(false);
+                comment.setBool(false);
+                comment.setLikeCnt(comment.getLikeCnt() - 1); // 좋아요를 삭제했으면 좋아요 수 감소
+                ApiResponseDto apiResponseDto = new ApiResponseDto("댓글 좋아요가 삭제되었습니다.", HttpStatus.OK.value());
+                return new ResponseEntity<>(apiResponseDto, HttpStatus.OK);
             } else {
-                comment.setBool(false); // 사용자가 좋아요 설정: true면 좋아요 생성, false: 좋아요 삭제
-                comment.setLikeCnt(comment.getLikeCnt() - 1); // 좋아요를 했으면 좋아요 수 감소
+                throw new Exception("작성자만 삭제할 수 있습니다.");
             }
-            commentRepository.deleteById(id);
-            ApiResponseDto apiResponseDto = new ApiResponseDto("댓글 좋아요가 삭제되었습니다.", HttpStatus.OK.value());
-            return new ResponseEntity<>(apiResponseDto, HttpStatus.OK);
         } catch (Exception e) {
             return ResponseEntity.badRequest().body(new ApiResponseDto(e.getMessage(), HttpStatus.OK.value()));
         }
